+++ src/src/main/java/Calculator/Equation/Operators/IOperator.java
@@ -1,5 +1,5 @@
 package Calculator.Equation.Operators;
 
 public interface IOperator {
-    public Integer Operate(Integer left , Integer right);
+    public Float Operate(Float left , Float right);
 }
